Crate spin_expressions

Re-exports

pub use provider::Provider;

pub use async_trait;

Modules

provider

Structs

Key

A variable key

PreparedResolver

A resolver who has resolved all variables.

ProviderResolver

A Resolver which is extended by Providers.

Resolver

A variable resolver.

Template

Template represents a simple string template that allows expressions in double curly braces, similar to Mustache or Liquid.

Enums

Error

A variable resolution error.

Type Aliases

Result

SharedPreparedResolver

A ProviderResolver that can be shared.